Re-use same outs & chosen mix outs across tx construction attempts
Upon completing round 1 of tx construction, the client determines
if the fee used in the tx is > result of calculate_fee(). If not,
it re-attempts the tx using the result of calculate_fee(). If the
client does not re-use the same set of mix_outs from the prior
attempt for each out, then the size of the transaction likely
ends up being a little different (because transactions reference
mix outs by on optimized offset, which varies in size depending
on the set of mix outs), even though the attempt is targeting
a fee for a transaction of an expected size. This can result in
the transaction including a fee that is *not* the result of
calculate_fee(final_tx). Thus, this change provides a way for
the client to re-use mix outs across attempts, so that the final
fee is equivalent ot the result of calculate_fee. This is now
implemented the same way as wallet2 (which also re-uses the
same set of decoys across attempts).

Bonus: it avoids trips to the server to get new decoys every
tx attempt as well.